+++ apps/api/src/app/auth/generate-authentication-options.dto.ts
@@ -1,0 +1,6 @@
+import { IsUUID } from 'class-validator';
+
+export class GenerateAuthenticationOptionsDto {
+  @IsUUID()
+  deviceId: string;
+}