+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Hardened the validation of the device id in the biometric authentication
+
 ### Fixed
 
 - Fixed the net performance percentage of date ranges in the portfolio performance calculation by including the gross performance at the start date